rtm
Former-commit-id: 36e7d66d5a5162cbf3e71761bd85d0c18882119d

+++ user/fangg/Mstackimg.c
@@ -34,7 +34,7 @@ int main(int argc, char* argv[])
 
     int ii, ix, iz, ix0, iz0, nin;
     int snx, snz, lnx, lnz;
-    int sox, soz, lox, loz, sdx, sdz, ldx, ldz; 
+    float sox, soz, lox, loz, sdx, sdz, ldx, ldz; 
     sf_file Fins, Fin, Fout;
     
     const char **filelist;
